error: fix deprecated usage of sys_errlist
The array sys_errlist is marked as deprecated in recent versions
of glibc.  The code already attempted to use strerror_r but it
was not being used because of missing include of config.h.

But the unused code to use strerror_r had another issue because
there are two versions of strerror_r in glibc. Use the defined
feature test macro (man strerror_r) to handle both versions.